+++ viewith-server/src/main/java/com/puremadeleine/viewith/aware/SpringProxyAware.java
@@ -1,4 +1,4 @@
-package com.puremadeleine.viewith;
+package com.puremadeleine.viewith.aware;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
